scripts/requirements.txt
psycopg2-binary
pandas
numpy
scipy
